# cop-sys-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For COP-SYSTEM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/department": {
            "post": {
                "tags": [
                    "Department"
                ],
                "summary": "Add a new department",
                "operationId": "355ead454e6307cbd2ec69db8d8ebb27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDepartmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Department added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepartmentsResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/department/{department}": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Get a department by ID",
                "operationId": "2a5510ab8e240bfd57e0d8cb8dfea659",
                "parameters": [
                    {
                        "name": "department",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepartmentsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Department not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Department"
                ],
                "summary": "Update an existing department",
                "operationId": "15c11f6162de545878cb194a38d77a2d",
                "parameters": [
                    {
                        "name": "department",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDepartmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Department updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepartmentsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Department not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Department"
                ],
                "summary": "Remove a department",
                "operationId": "1b906308f133bc3ad5d94df22d5974fb",
                "parameters": [
                    {
                        "name": "department",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department deleted successfully"
                    },
                    "404": {
                        "description": "Department not found"
                    }
                }
            }
        },
...
        "/api/v1/equipment": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Get all equipment",
                "operationId": "197bcfcae3e8e75c109a7d52c3b35faa",
                "responses": {
                    "200": {
                        "description": "Equipments list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EquipmentsResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Add a new equipment",
                "operationId": "e0fc0ad9252b6a72249541042d967768",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEquipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Equipment added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentsResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/equipment/{equipment}": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Get an equipment by ID",
                "operationId": "716a2624e922650b8e0e652ae98ea6e0",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipment retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipment not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Update an existing equipment",
                "operationId": "825f0b3d2d88c59c65db358146a24a2a",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEquipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Equipment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipment not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Remove an equipment",
                "operationId": "9452e643b9cf83bf89c7098484a2f4c8",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipment deleted successfully"
                    },
                    "404": {
                        "description": "Equipment not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateEquipmentRequest": {
                "required": [
                    "name",
                    "quantity"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateEquipmentRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "EquipmentsResource": {
                "properties": {
                    "equipment_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CreateDepartmentRequest": {
                "required": [
                    "department_name",
                    "department_logo",
                    "description"
                ],
                "properties": {
                    "department_name": {
                        "type": "string"
                    },
                    "department_logo": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateDepartmentRequest": {
                "properties": {
                    "department_name": {
                        "type": "string"
                    },
                    "department_logo": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DepartmentsResource": {
                "properties": {
                    "department_id": {
                        "type": "integer"
                    },
                    "department_name": {
                        "type": "string"
                    },
                    "department_logo": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}